move graphql response assertions to json lib
